+++ financial-connections/src/test/snapshots/images/com.stripe.android.financialconnections.screenshottests_PaparazziSampleScreenshotTest_preview_tests[com.stripe.android.financialconnections.features.linkaccountpicker_null_LinkAccountPickerPane_LinkAccountPickerScreenPreview_0_null_0,NEXUS_5].png
